pyproject.toml
[project]
name = "rovi-control-board"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "influxdb-client>=1.49.0",
    "pygame>=2.6.1",
    "pyserial>=3.5",
    "python-dotenv>=1.2.1",
    "rosmaster-lib",
    "PyYAML>=6.0.2",
    "numpy>=2.3.4",
]

[tool.uv.sources]
rosmaster-lib = { git = "https://github.com/RobLibs/Rosmaster_Lib" }
